remove redundant files and fix regression
