--- docs/src/modules/components/HookApiPage.js
@@ -1,178 +1,0 @@
-/* eslint-disable react/no-danger */
-import * as React from 'react';
-import PropTypes from 'prop-types';
-import { exactProp } from '@mui/utils';
-import Typography from '@mui/material/Typography';
-import { useTranslate, useUserLanguage } from 'docs/src/modules/utils/i18n';
-import PropertiesTable from 'docs/src/modules/components/PropertiesTable';
-import HighlightedCode from 'docs/src/modules/components/HighlightedCode';
-import MarkdownElement from 'docs/src/modules/components/MarkdownElement';
-import AppLayoutDocs from 'docs/src/modules/components/AppLayoutDocs';
-import Ad from 'docs/src/modules/components/Ad';
-
-export function getTranslatedHeader(t, header) {
-  const translations = {
-    demos: t('api-docs.demos'),
-    import: t('api-docs.import'),
-    'hook-name': t('api-docs.hookName'),
-    parameters: t('api-docs.parameters'),
-    'return-value': t('api-docs.returnValue'),
-  };
-
-  // TODO Drop runtime type-checking once we type-check this file
-  if (!translations.hasOwnProperty(header)) {
-    throw new TypeError(
-      `Unable to translate header '${header}'. Did you mean one of '${Object.keys(
-        translations,
-      ).join("', '")}'`,
-    );
-  }
-
-  return translations[header] || header;
-}
-
-function Heading(props) {
-  const { hash, level: Level = 'h2' } = props;
-  const t = useTranslate();
-
-  return (
-    <Level id={hash}>
-      {getTranslatedHeader(t, hash)}
-      <a aria-labelledby={hash} className="anchor-link" href={`#${hash}`} tabIndex={-1}>
-        <svg>
-          <use xlinkHref="#anchor-link-icon" />
-        </svg>
-      </a>
-    </Level>
-  );
-}
-
-Heading.propTypes = {
-  hash: PropTypes.string.isRequired,
-  level: PropTypes.string,
-};
-
-const nonDefaultHooks = ['useFormControlUnstyledContext'];
-
-export default function ApiPage(props) {
-  const { descriptions, disableAd = false, pageContent } = props;
-  const t = useTranslate();
-  const userLanguage = useUserLanguage();
-
-  const { demos, filename, inheritance, name: hookName, parameters, returnValue } = pageContent;
-
-  const {
-    hookDescription,
-    hookDescriptionToc = [],
-    parametersDescriptions,
-    returnValueDescriptions,
-  } = descriptions[userLanguage];
-  const description = t('api-docs.hooksPageDescription').replace(/{{name}}/, hookName);
-
-  const source = filename.replace(
-    /\/packages\/mui(-(.+?))?\/src/,
-    (match, dash, pkg) => `@mui/${pkg}`,
-  );
-
-  // Prefer linking the .tsx or .d.ts for the "Edit this page" link.
-  const apiSourceLocation = filename.replace('.js', '.d.ts');
-
-  function createTocEntry(sectionName) {
-    return {
-      text: getTranslatedHeader(t, sectionName),
-      hash: sectionName,
-      children: [
-        ...(sectionName === 'props' && inheritance
-          ? [{ text: t('api-docs.inheritance'), hash: 'inheritance', children: [] }]
-          : []),
-      ],
-    };
-  }
-
-  const toc = [
-    createTocEntry('demos'),
-    createTocEntry('import'),
-    ...hookDescriptionToc,
-    createTocEntry('parameters'),
-    createTocEntry('return-value'),
-  ].filter(Boolean);
-
-  return (
-    <AppLayoutDocs
-      description={description}
-      disableAd={disableAd}
-      disableToc={false}
-      location={apiSourceLocation}
-      title={`${hookName} API`}
-      toc={toc}
-    >
-      <MarkdownElement>
-        <h1>{hookName} API</h1>
-        <Typography
-          variant="h5"
-          component="p"
-          className={`description${disableAd ? '' : ' ad'}`}
-          gutterBottom
-        >
-          {description}
-          {disableAd ? null : <Ad />}
-        </Typography>
-        <Heading hash="demos" />
-        <div
-          className="MuiCallout-root MuiCallout-info"
-          dangerouslySetInnerHTML={{
-            __html: `<p>For examples and details on the usage of this React hook, visit the demo pages:</p>
-              ${demos}`,
-          }}
-        />
-        <Heading hash="import" />
-        <HighlightedCode
-          code={`
-import ${nonDefaultHooks.includes(hookName) ? `{ ${hookName} }` : hookName} from '${source
-            .split('/')
-            .slice(0, -1)
-            .join('/')}';`}
-          language="jsx"
-        />
-        {/* TODO: Add this once the hooks are in dedicated folders */}
-        {/* <span dangerouslySetInnerHTML={{ __html: t('api-docs.importDifference') }} /> */}
-        {hookDescription ? (
-          <React.Fragment>
-            <br />
-            <br />
-            <span
-              dangerouslySetInnerHTML={{
-                __html: hookDescription,
-              }}
-            />
-          </React.Fragment>
-        ) : null}
-        <Heading hash="parameters" />
-        <PropertiesTable properties={parameters} propertiesDescriptions={parametersDescriptions} />
-        <Heading hash="return-value" />
-        <PropertiesTable
-          showOptionalAbbr
-          properties={returnValue}
-          propertiesDescriptions={returnValueDescriptions}
-        />
-        <br />
-        {/* TODO: Add section for the hook output type */}
-      </MarkdownElement>
-      <svg style={{ display: 'none' }} xmlns="http://www.w3.org/2000/svg">
-        <symbol id="anchor-link-icon" viewBox="0 0 16 16">
-          <path d="M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z" />
-        </symbol>
-      </svg>
-    </AppLayoutDocs>
-  );
-}
-
-ApiPage.propTypes = {
-  descriptions: PropTypes.object.isRequired,
-  disableAd: PropTypes.bool,
-  pageContent: PropTypes.object.isRequired,
-};
-
-if (process.env.NODE_ENV !== 'production') {
-  ApiPage.propTypes = exactProp(ApiPage.propTypes);
-}
